refactor(gateway): GAR-478 — Q10.d extract build_slack_channels to bootstrap/slack.rs
Extracts build_slack_channels (166 LOC) from bootstrap/mod.rs into a new
sibling module bootstrap/slack.rs. Re-exports via pub use at the original
path — zero behaviour change. bootstrap/mod.rs: 2044 → 1878 lines (−166).

Plan 0166. Parent epic: GAR-440.
